+++ src/main/java/net/fabricmc/loom/task/LoomTasks.java
@@ -57,6 +57,13 @@ public final class LoomTasks {
 		tasks.register("downloadAssets", DownloadAssetsTask.class, t -> t.setDescription("Downloads required assets for Fabric."));
 		tasks.register("remapSourcesJar", RemapSourcesJarTask.class, t -> t.setDescription("Remaps the project sources jar to intermediary names."));
 
+		tasks.getByName("check").dependsOn(
+				tasks.register("validateAccessWidener", ValidateAccessWidenerTask.class, t -> {
+					t.setDescription("Validate that all the rules in the access widener against the minecraft jar");
+					t.setGroup("verification");
+				})
+		);
+
 		registerIDETasks(tasks);
 		registerRunTasks(tasks, project);
 		registerDecompileTasks(tasks, project);
